Refactor fork resolution
Move fork-resolution code to a separate module. Move support
functions along with it. Replace utility calls with calls to
poet2_utils.
